requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
attrs=23.1.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
blinker=1.6.2=pypi_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.05.30=haa95532_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.0=pypi_0
cryptography=41.0.2=py38h31511bf_0
cuda-cccl=12.2.140=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.2.140=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
entrypoints=0.4=pypi_0
filelock=3.9.0=py38haa95532_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.34=pypi_0
idna=3.4=py38haa95532_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kiwisolver=1.4.5=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.3.141=0
libcurand-dev=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.6=py38hf11a4ad_1
mkl_random=1.2.2=py38hf11a4ad_1
mpmath=1.3.0=py38haa95532_0
networkx=3.1=py38haa95532_0
numpy=1.24.4=pypi_0
numpy-base=1.24.3=py38h8a87ada_1
opencv-python=4.7.0.68=pypi_0
openssl=1.1.1v=h2bbff1b_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
pyarrow=13.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.1b0=pypi_0
pygments=2.16.1=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.8_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
referencing=0.30.2=pypi_0
requests=2.31.0=py38haa95532_0
rich=13.5.2=pypi_0
rpds-py=0.10.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
semver=3.0.1=pypi_0
setuptools=68.0.0=py38haa95532_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
streamlit=1.11.0=pypi_0
sympy=1.11.1=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
typing_extensions=4.7.1=py38haa95532_0
tzdata=2023.3=pypi_0
tzlocal=5.0.1=pypi_0
urllib3=2.0.4=pypi_0
validators=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
